+++ DOCS/CHANGELOG.md
@@ -27,6 +27,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 - structs:
   - namespaces: `Arch`, `ByteSearch`, `Char`, `Env`, `Mem`, `Ptr`, `Str`.
   - `CompressionMode`, `EncodingMode`, `Pnm`.
+  - standalone error types: `ErrorNotImplemented`, `ErrorNotSupported`, `ErrorElementNotFound`, `ErrorInvalidAxisLength`, `ErrorKeyAlreadyExists`, `ErrorMismatchedDimensions`, `ErrorMismatchedIndices`, `ErrorMismatchedLength`, `ErrorNodeEmpty`, `ErrorNodeLinkNotSet`, `ErrorNodeLinkNotUnique`, `ErrorNotEnoughElements`, `ErrorNotEnoughSpace`, `ErrorOutOfBounds`, `ErrorOverflow`, `ErrorPartiallyAdded`.
   - `False`, `True`, `UnitBi`, `UnitSi`.
   - `HasherPengy`.
   - `Lgc16`.